+++ payment-system/paymentservice/src/test/kotlin/org/example/paymentservice/payment/test/R2DBCPaymentDatabaseHelper.kt
@@ -43,7 +43,7 @@ class R2DBCPaymentDatabaseHelper(
                                 sellerId = result["seller_id"] as Long,
                                 orderId = result["order_id"] as String,
                                 productId = result["product_id"] as Long,
-                                amount = result["amount"] as BigDecimal,
+                                amount = (result["amount"] as BigDecimal).toLong(),
                                 paymentStatus = PaymentStatus.get(result["payment_order_status"] as String),
                                 isLedgerUpdated = ((result["ledger_updated"] as Byte).toInt() == 1),
                                 isWalletUpdated = ((result["wallet_updated"] as Byte).toInt() == 1),
